refactor: move common utils
