+++ packages/components/progress/progress.tsx
@@ -8,7 +8,7 @@ import {
   ErrorIcon as TdErrorIcon,
 } from 'tdesign-icons-vue-next';
 
-import { getBackgroundColor } from '../utils/helper';
+import { getBackgroundColor } from '../../utils/helper';
 import { PRO_THEME, CIRCLE_SIZE, CIRCLE_SIZE_PX, STATUS_ICON, CIRCLE_FONT_SIZE_RATIO } from './constants';
 import props from './props';
 import { usePrefixClass } from '../hooks/useConfig';
