+++ packages/components/utils/dom.ts
@@ -8,7 +8,7 @@ import raf from 'raf';
 import { isString } from 'lodash-es';
 import { isFunction } from 'lodash-es';
 import { isArray } from 'lodash-es';
-import { easeInOutCubic, EasingFunction } from './easing';
+import { easeInOutCubic, EasingFunction } from '../../utils/easing';
 import { ScrollContainer, ScrollContainerElement } from '../common';
 
 export const isServer = typeof window === 'undefined';
