+++ packages/components/utils/event.ts
@@ -2,7 +2,7 @@ import { ComponentPublicInstance } from 'vue';
 import { isFunction } from 'lodash-es';
 import { isString } from 'lodash-es';
 
-import { getPropsApiByEvent } from './helper';
+import { getPropsApiByEvent } from '../../utils/helper';
 export type EmitEventName = { event: string; method: string } | string;
 
 /**
